+++ Wabbajack.DTOs/Game/GameRegistry.cs
@@ -492,6 +492,23 @@ public static class GameRegistry
                 },
                 MainExecutable = "nw.exe".ToRelativePath()
             }
+        },
+        {
+            Game.Valheim, new GameMetaData
+            {
+                Game = Game.Valheim,
+                SteamIDs = new[] { 892970 },
+                MO2Name = "Valheim",
+                MO2ArchiveName = "valheim",
+                NexusName = "valheim",
+                NexusGameId = 3667,
+                IsGenericMO2Plugin = true,
+                RequiredFiles = new []
+                {
+                    "valheim.exe".ToRelativePath()
+                },
+                MainExecutable = "valheim.exe".ToRelativePath()
+            }
         }
     };