+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### Version - 3.0.x.x - TBD
 * Game support:
   * Added support for Valheim
+  * Added Cyberpunk 2077 Epic Games Store support
 
 #### Version - 3.0.4.1 - 11/17/2022
 * Fix recursive writing of meta files for meta files