+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ### Changelog
 
 #### Version - 3.0.x.x - TBD
+
+* Fix Website Links
 * Game support:
   * Added support for Valheim
   * Added Cyberpunk 2077 Epic Games Store support